Fix warnings

+++ Geo/GEdge.cpp
@@ -778,7 +778,7 @@ bool GEdge::reorder(const int elementType, const std::vector<std::size_t> &order
 
     for(std::vector<std::size_t>::const_iterator it = ordering.begin();
         it != ordering.end(); ++it) {
-      if(*it < 0 || *it >= static_cast<int>(lines.size())) return false;
+      if(*it < 0 || *it >= lines.size()) return false;
     }
 
     std::vector<MLine *> newLinesOrder(lines.size());
